+++ packages/app-lib/src/api/instance/synced_servers/operations.rs
@@ -1,154 +1,34 @@
-use super::synced_options::{
+use super::super::synced_options::{
     CheckpointStatus, checkpoint, detach_link, ensure_link, finish_checkpoint,
     instance_dir, instance_is_running, instance_option_enabled,
-    instance_option_supported, nbt_from_bytes, nbt_to_bytes, read_nbt_file,
-    safe_instance_id, sha1_bytes, sha1_file, sync_files_are_protected,
-    synced_options_path,
+    instance_option_supported, sha1_bytes, sha1_file, sync_files_are_protected,
 };
-use crate::state::{CachedEntry, InstanceLink, InstanceMetadata, SyncedOption};
-use crate::util::fetch::{DownloadMeta, DownloadReason, fetch_mirrors};
+use crate::state::{InstanceMetadata, SyncedOption};
 use crate::{ErrorKind, State};
-use async_zip::base::read::seek::ZipFileReader;
-use quartz_nbt::{NbtCompound, NbtList, NbtTag};
-use serde::{Deserialize, Serialize};
-use sqlx::{Sqlite, Transaction};
-use std::collections::HashSet;
-use std::io::Cursor;
-use std::path::{Path, PathBuf};
+use quartz_nbt::NbtCompound;
 use uuid::Uuid;
 
-const SERVERS_FILE: &str = "servers.dat";
+use super::SERVERS_FILE;
+use super::codec::{
+    read_servers, server_address, server_hidden, server_identity_address,
+    server_identity_name, servers_to_bytes, update_server_data, write_servers,
+};
+use super::modpack::{
+    is_modpack_link, pack_state_exists, pack_state_matches_link,
+    reconstruct_modpack_servers,
+};
+use super::storage::{
+    begin_server_checkpoint, canonical_exists, commit_server_state,
+    generated_path, load_local, load_projection_entries, read_canonical,
+    server_revision, write_local,
+};
+use super::types::{
+    CanonicalServer, DesyncServerMode, LocalServer, ProjectionEntry,
+    ProjectionOwner, ServerRecord, ServerSource, SyncedServer,
+};
+use std::collections::HashSet;
 
-#[derive(Clone, Copy, Debug, Eq, PartialEq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum ServerSource {
-    UserSynced,
-    Modpack,
-    LinkedServerProject,
-    LocalDesynced,
-}
-
-impl ServerSource {
-    fn as_str(self) -> &'static str {
-        match self {
-            Self::UserSynced => "user_synced",
-            Self::Modpack => "modpack",
-            Self::LinkedServerProject => "linked_server_project",
-            Self::LocalDesynced => "local_desynced",
-        }
-    }
-
-    fn from_str(value: &str) -> Option<Self> {
-        match value {
-            "user_synced" => Some(Self::UserSynced),
-            "modpack" => Some(Self::Modpack),
-            "linked_server_project" => Some(Self::LinkedServerProject),
-            "local_desynced" => Some(Self::LocalDesynced),
-            _ => None,
-        }
-    }
-}
-
-#[derive(Clone, Copy, Debug, Eq, PartialEq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum DesyncServerMode {
-    KeepInOtherInstances,
-    RemoveFromOtherInstances,
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize)]
-pub struct SyncedServer {
-    pub id: String,
-    pub name: String,
-    pub address: String,
-    pub accept_textures: Option<bool>,
-}
-
-#[derive(Clone, Debug)]
-pub(crate) struct ServerRecord {
-    pub id: String,
-    pub source: ServerSource,
-    pub data: NbtCompound,
-}
-
-impl ServerRecord {
-    pub fn name(&self) -> String {
-        self.data
-            .get::<_, &str>("name")
-            .unwrap_or_default()
-            .to_string()
-    }
-
-    pub fn address(&self) -> String {
-        self.data
-            .get::<_, &str>("ip")
-            .unwrap_or_default()
-            .to_string()
-    }
-
-    pub fn icon(&self) -> Option<String> {
-        self.data.get::<_, &str>("icon").ok().map(ToOwned::to_owned)
-    }
-
-    pub fn hidden(&self) -> bool {
-        self.data.get::<_, i8>("hidden").unwrap_or(0) != 0
-    }
-
-    pub fn accept_textures(&self) -> Option<bool> {
-        self.data
-            .get::<_, i8>("acceptTextures")
-            .ok()
-            .map(|value| value != 0)
-    }
-}
-
-#[derive(Clone)]
-struct CanonicalServer {
-    id: String,
-    data: NbtCompound,
-}
-
-#[derive(Clone)]
-struct LocalServer {
-    id: String,
-    source: ServerSource,
-    excluded_synced_server_id: Option<String>,
-    data: NbtCompound,
-    position: i64,
-}
-
-#[derive(Clone)]
-struct ProjectionEntry {
-    id: String,
-    owner: ProjectionOwner,
-    data: NbtCompound,
-    position: i64,
-}
-
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-enum ProjectionOwner {
-    Synced,
-    Instance,
-}
-
-impl ProjectionOwner {
-    fn as_str(self) -> &'static str {
-        match self {
-            Self::Synced => "synced",
-            Self::Instance => "instance",
-        }
-    }
-
-    fn from_str(value: &str) -> Option<Self> {
-        match value {
-            "synced" => Some(Self::Synced),
-            "instance" => Some(Self::Instance),
-            _ => None,
-        }
-    }
-}
-
-pub(super) async fn seed_servers(
+pub(crate) async fn seed_servers(
     metadata: &InstanceMetadata,
     state: &State,
 ) -> crate::Result<()> {
@@ -195,7 +75,7 @@ pub(super) async fn seed_servers(
     regenerate_servers(state).await
 }
 
-pub(super) async fn merge_servers_from_instance(
+pub(crate) async fn merge_servers_from_instance(
     metadata: &InstanceMetadata,
     state: &State,
 ) -> crate::Result<()> {
@@ -253,7 +133,7 @@ pub(super) async fn merge_servers_from_instance(
     Ok(())
 }
 
-pub(super) async fn ensure_servers(
+pub(crate) async fn ensure_servers(
     metadata: &InstanceMetadata,
     state: &State,
 ) -> crate::Result<()> {
@@ -286,7 +166,7 @@ pub(super) async fn ensure_servers(
     compose_instance(metadata, state).await
 }
 
-pub(super) async fn detach_servers(
+pub(crate) async fn detach_servers(
     metadata: &InstanceMetadata,
     state: &State,
 ) -> crate::Result<()> {
@@ -295,192 +175,7 @@ pub(super) async fn detach_servers(
     detach_link(&generated, &local).await
 }
 
-pub async fn capture_modpack_servers(instance_id: &str) -> crate::Result<()> {
-    let state = State::get().await?;
-    let _guard = state.lock_synced_options().await;
-    let metadata = crate::state::get_instance(instance_id, &state.pool)
-        .await?
-        .ok_or_else(|| ErrorKind::InputError("Unknown instance".to_string()))?;
-    let path = instance_dir(&metadata, &state).join(SERVERS_FILE);
-    let servers = read_servers(&path).await?;
-    replace_modpack_servers(&metadata, servers, &state).await?;
-    if effective(&metadata, &state).await? {
-        compose_instance(&metadata, &state).await?;
-    }
-    Ok(())
-}
-
-pub async fn clear_modpack_servers(instance_id: &str) -> crate::Result<()> {
-    let state = State::get().await?;
-    let _guard = state.lock_synced_options().await;
-    let metadata = crate::state::get_instance(instance_id, &state.pool)
-        .await?
-        .ok_or_else(|| ErrorKind::InputError("Unknown instance".to_string()))?;
-    replace_modpack_servers(&metadata, Vec::new(), &state).await?;
-    if effective(&metadata, &state).await? {
-        compose_instance(&metadata, &state).await?;
-    }
-    Ok(())
-}
-
-async fn replace_modpack_servers(
-    metadata: &InstanceMetadata,
-    servers: Vec<NbtCompound>,
-    state: &State,
-) -> crate::Result<()> {
-    let mut local = servers
-        .into_iter()
-        .enumerate()
-        .map(|(position, data)| LocalServer {
-            id: Uuid::new_v4().to_string(),
-            source: ServerSource::Modpack,
-            excluded_synced_server_id: None,
-            data,
-            position: position as i64,
-        })
-        .collect::<Vec<_>>();
-    let mut retained = load_local(&metadata.instance.id, state)
-        .await?
-        .into_iter()
-        .filter(|server| server.source != ServerSource::Modpack)
-        .collect::<Vec<_>>();
-    for server in &mut retained {
-        server.position = local.len() as i64;
-        local.push(server.clone());
-    }
-    let mut tx = state.pool.begin().await?;
-    write_local_rows(&mut tx, &metadata.instance.id, &local).await?;
-    let version_id = modpack_version_id(&metadata.link);
-    sqlx::query!(
-        "
-		INSERT INTO instance_server_pack_state (instance_id, version_id)
-		VALUES (?, ?)
-		ON CONFLICT(instance_id) DO UPDATE SET
-			version_id = excluded.version_id
-		",
-        metadata.instance.id,
-        version_id,
-    )
-    .execute(&mut *tx)
-    .await?;
-    tx.commit().await?;
-    Ok(())
-}
-
-async fn pack_state_matches_link(
-    metadata: &InstanceMetadata,
-    state: &State,
-) -> crate::Result<bool> {
-    let row = sqlx::query!(
-        "
-		SELECT version_id
-		FROM instance_server_pack_state
-		WHERE instance_id = ?
-		",
-        metadata.instance.id,
-    )
-    .fetch_optional(&state.pool)
-    .await?;
-    Ok(row.is_some_and(|row| {
-        row.version_id.as_deref() == modpack_version_id(&metadata.link)
-    }))
-}
-
-async fn pack_state_exists(
-    instance_id: &str,
-    state: &State,
-) -> crate::Result<bool> {
-    Ok(sqlx::query_scalar!(
-        r#"
-		SELECT EXISTS(
-			SELECT 1 FROM instance_server_pack_state WHERE instance_id = ?
-		) AS "exists!: bool"
-		"#,
-        instance_id,
-    )
-    .fetch_one(&state.pool)
-    .await?)
-}
-
-async fn reconstruct_modpack_servers(
-    metadata: &InstanceMetadata,
-    state: &State,
-) -> crate::Result<()> {
-    let version_id = modpack_version_id(&metadata.link).ok_or_else(|| {
-        ErrorKind::InputError(
-            "This modpack does not have a recoverable Modrinth version."
-                .to_string(),
-        )
-    })?;
-    let version = CachedEntry::get_version(
-        version_id,
-        None,
-        &state.pool,
-        &state.api_semaphore,
-    )
-    .await?
-    .ok_or_else(|| {
-        ErrorKind::InputError(format!(
-            "Modpack version {version_id} could not be found"
-        ))
-    })?;
-    let primary_file = version
-        .files
-        .iter()
-        .find(|file| file.primary)
-        .or_else(|| version.files.first())
-        .ok_or_else(|| {
-            ErrorKind::InputError(format!(
-                "Modpack version {version_id} has no downloadable file"
-            ))
-        })?;
-    let download_meta = DownloadMeta {
-        reason: DownloadReason::Modpack,
-        game_version: metadata.applied_content_set.game_version.clone(),
-        loader: metadata.applied_content_set.loader.as_str().to_string(),
-        dependent_on: Some(version_id.to_string()),
-    };
-    let mrpack = fetch_mirrors(
-        &[&primary_file.url],
-        primary_file.hashes.get("sha1").map(String::as_str),
-        Some(&download_meta),
-        None,
-        &state.api_semaphore,
-        &state.pool,
-    )
-    .await?;
-    let mut archive = ZipFileReader::with_tokio(Cursor::new(&mrpack)).await?;
-    let mut selected = None;
-    for (index, entry) in archive.file().entries().iter().enumerate() {
-        let Ok(filename) = entry.filename().as_str() else {
-            continue;
-        };
-        let priority = if filename == "client-overrides/servers.dat" {
-            2
-        } else if filename == "overrides/servers.dat" {
-            1
-        } else {
-            continue;
-        };
-        if selected
-            .is_none_or(|(selected_priority, _)| priority >= selected_priority)
-        {
-            selected = Some((priority, index));
-        }
-    }
-
-    let servers = if let Some((_, index)) = selected {
-        let mut bytes = Vec::new();
-        let mut reader = archive.reader_with_entry(index).await?;
-        reader.read_to_end_checked(&mut bytes).await?;
-        servers_from_bytes(bytes)?
-    } else {
-        Vec::new()
-    };
-    replace_modpack_servers(metadata, servers, state).await
-}
-
-pub(super) async fn reconcile_servers(
+pub(crate) async fn reconcile_servers(
     metadata: &InstanceMetadata,
     state: &State,
 ) -> crate::Result<()> {
@@ -1005,7 +700,7 @@ pub async fn remove_synced_server(server_id: &str) -> crate::Result<()> {
     regenerate_servers(&state).await
 }
 
-async fn compose_instance(
+pub(crate) async fn compose_instance(
     metadata: &InstanceMetadata,
     state: &State,
 ) -> crate::Result<()> {
@@ -1089,7 +784,7 @@ async fn regenerate_servers(state: &State) -> crate::Result<()> {
     Ok(())
 }
 
-async fn effective(
+pub(crate) async fn effective(
     metadata: &InstanceMetadata,
     state: &State,
 ) -> crate::Result<bool> {
@@ -1126,475 +821,4 @@ async fn participating(
         state,
     )
     .await)
-}
-
-pub(super) async fn canonical_exists(state: &State) -> crate::Result<bool> {
-    canonical_initialized(state).await
-}
-
-async fn canonical_initialized(state: &State) -> crate::Result<bool> {
-    Ok(sqlx::query_scalar!(
-        r#"
-		SELECT EXISTS(
-			SELECT 1 FROM synced_server_state WHERE singleton = 1
-		) AS "initialized!: bool"
-		"#,
-    )
-    .fetch_one(&state.pool)
-    .await?)
-}
-
-async fn read_canonical(state: &State) -> crate::Result<Vec<CanonicalServer>> {
-    let rows = sqlx::query!(
-        "
-		SELECT id, nbt
-		FROM synced_servers
-		ORDER BY position
-		",
-    )
-    .fetch_all(&state.pool)
-    .await?;
-    rows.into_iter()
-        .map(|row| {
-            Ok(CanonicalServer {
-                id: row.id,
-                data: nbt_from_bytes(row.nbt)?,
-            })
-        })
-        .collect()
-}
-
-async fn commit_server_state(
-    canonical: Option<&[CanonicalServer]>,
-    local: Option<(&str, &[LocalServer])>,
-    state: &State,
-) -> crate::Result<bool> {
-    let mut tx = state.pool.begin().await?;
-    let mut canonical_changed = false;
-    if let Some(canonical) = canonical {
-        canonical_changed = write_canonical_rows(&mut tx, canonical).await?;
-        sqlx::query!(
-            "
-			INSERT INTO synced_server_state (singleton, revision)
-			VALUES (1, 0)
-			ON CONFLICT(singleton) DO NOTHING
-			",
-        )
-        .execute(&mut *tx)
-        .await?;
-        if canonical_changed {
-            sqlx::query!(
-                "
-				UPDATE synced_server_state
-				SET revision = revision + 1
-				WHERE singleton = 1
-				",
-            )
-            .execute(&mut *tx)
-            .await?;
-        }
-    }
-    if let Some((instance_id, local)) = local {
-        write_local_rows(&mut tx, instance_id, local).await?;
-    }
-    tx.commit().await?;
-    Ok(canonical_changed)
-}
-
-async fn write_canonical_rows(
-    tx: &mut Transaction<'_, Sqlite>,
-    servers: &[CanonicalServer],
-) -> crate::Result<bool> {
-    let current = sqlx::query!(
-        "
-		SELECT id, position, nbt
-		FROM synced_servers
-		ORDER BY position
-		",
-    )
-    .fetch_all(&mut **tx)
-    .await?;
-    let mut desired = Vec::with_capacity(servers.len());
-    for (position, server) in servers.iter().enumerate() {
-        desired.push((
-            server.id.as_str(),
-            position as i64,
-            nbt_to_bytes(&server.data)?,
-        ));
-    }
-    if current.len() == desired.len()
-        && current.iter().zip(&desired).all(|(current, desired)| {
-            current.id == desired.0
-                && current.position == desired.1
-                && current.nbt == desired.2
-        })
-    {
-        return Ok(false);
-    }
-
-    let desired_ids =
-        desired.iter().map(|(id, _, _)| *id).collect::<HashSet<_>>();
-    for row in &current {
-        if !desired_ids.contains(row.id.as_str()) {
-            sqlx::query!("DELETE FROM synced_servers WHERE id = ?", row.id)
-                .execute(&mut **tx)
-                .await?;
-        }
-    }
-
-    let position_offset = current
-        .iter()
-        .map(|row| row.position)
-        .max()
-        .unwrap_or(0)
-        .max(servers.len() as i64)
-        .saturating_add(1);
-    sqlx::query!(
-        "UPDATE synced_servers SET position = position + ?",
-        position_offset,
-    )
-    .execute(&mut **tx)
-    .await?;
-
-    for (id, position, nbt) in desired {
-        sqlx::query!(
-            "
-			INSERT INTO synced_servers (id, position, nbt)
-			VALUES (?, ?, ?)
-			ON CONFLICT(id) DO UPDATE SET
-				position = excluded.position,
-				nbt = excluded.nbt
-			",
-            id,
-            position,
-            nbt,
-        )
-        .execute(&mut **tx)
-        .await?;
-    }
-    Ok(true)
-}
-
-async fn server_revision(state: &State) -> crate::Result<i64> {
-    Ok(sqlx::query_scalar!(
-        r#"
-		SELECT revision AS "revision!: i64"
-		FROM synced_server_state
-		WHERE singleton = 1
-		"#,
-    )
-    .fetch_optional(&state.pool)
-    .await?
-    .unwrap_or(0))
-}
-
-pub(crate) async fn read_servers(
-    path: &Path,
-) -> crate::Result<Vec<NbtCompound>> {
-    if !path.exists() {
-        return Ok(Vec::new());
-    }
-    servers_from_root(read_nbt_file(path).await?)
-}
-
-fn servers_from_bytes(bytes: Vec<u8>) -> crate::Result<Vec<NbtCompound>> {
-    servers_from_root(nbt_from_bytes(bytes)?)
-}
-
-fn servers_from_root(root: NbtCompound) -> crate::Result<Vec<NbtCompound>> {
-    let list = root.get::<_, &NbtList>("servers").map_err(|_| {
-        ErrorKind::InputError(
-            "servers.dat does not contain a valid servers list".to_string(),
-        )
-    })?;
-    list.iter()
-        .map(|tag| match tag {
-            NbtTag::Compound(compound) => Ok(compound.clone()),
-            _ => Err(ErrorKind::InputError(
-                "servers.dat contains an invalid server entry".to_string(),
-            )
-            .into()),
-        })
-        .collect()
-}
-
-pub(crate) async fn write_servers(
-    path: &Path,
-    servers: &[NbtCompound],
-) -> crate::Result<()> {
-    if let Some(parent) = path.parent() {
-        crate::util::io::create_dir_all(parent).await?;
-    }
-    crate::util::io::write(path, servers_to_bytes(servers)?).await?;
-    Ok(())
-}
-
-fn servers_to_bytes(servers: &[NbtCompound]) -> crate::Result<Vec<u8>> {
-    let mut list = NbtList::new();
-    for server in servers {
-        list.push(server.clone());
-    }
-    let mut root = NbtCompound::new();
-    root.insert("servers", list);
-    nbt_to_bytes(&root)
-}
-
-pub(crate) fn server_data(
-    name: String,
-    address: String,
-    accept_textures: Option<bool>,
-) -> NbtCompound {
-    let mut server = NbtCompound::new();
-    server.insert("name", name);
-    server.insert("ip", address);
-    if let Some(accept_textures) = accept_textures {
-        server.insert("acceptTextures", i8::from(accept_textures));
-    }
-    server.insert("hidden", 0_i8);
-    server
-}
-
-pub(crate) fn update_server_data(
-    server: &mut NbtCompound,
-    name: String,
-    address: String,
-    accept_textures: Option<bool>,
-) {
-    server.insert("name", name);
-    server.insert("ip", address);
-    match accept_textures {
-        Some(value) => server.insert("acceptTextures", i8::from(value)),
-        None => {
-            server.inner_mut().remove("acceptTextures");
-        }
-    }
-}
-
-fn server_hidden(server: &NbtCompound) -> bool {
-    server.get::<_, i8>("hidden").unwrap_or(0) != 0
-}
-
-fn server_address(server: &NbtCompound) -> String {
-    server.get::<_, &str>("ip").unwrap_or_default().to_string()
-}
-
-fn server_identity_address(server: &NbtCompound) -> String {
-    server_address(server).trim().to_ascii_lowercase()
-}
-
-fn server_identity_name(server: &NbtCompound) -> String {
-    server
-        .get::<_, &str>("name")
-        .unwrap_or_default()
-        .trim()
-        .to_ascii_lowercase()
-}
-
-async fn load_local(
-    instance_id: &str,
-    state: &State,
-) -> crate::Result<Vec<LocalServer>> {
-    let rows = sqlx::query!(
-        "
-		SELECT id, source, excluded_synced_server_id, nbt, position
-		FROM instance_servers
-		WHERE instance_id = ?
-		ORDER BY position
-		",
-        instance_id,
-    )
-    .fetch_all(&state.pool)
-    .await?;
-    rows.into_iter()
-        .map(|row| {
-            Ok(LocalServer {
-                id: row.id,
-                source: ServerSource::from_str(&row.source).ok_or_else(
-                    || {
-                        ErrorKind::InputError(format!(
-                            "Unknown server source {}",
-                            row.source
-                        ))
-                    },
-                )?,
-                excluded_synced_server_id: row.excluded_synced_server_id,
-                data: nbt_from_bytes(row.nbt)?,
-                position: row.position,
-            })
-        })
-        .collect()
-}
-
-async fn write_local(
-    instance_id: &str,
-    servers: &[LocalServer],
-    state: &State,
-) -> crate::Result<()> {
-    commit_server_state(None, Some((instance_id, servers)), state).await?;
-    Ok(())
-}
-
-async fn write_local_rows(
-    tx: &mut Transaction<'_, Sqlite>,
-    instance_id: &str,
-    servers: &[LocalServer],
-) -> crate::Result<()> {
-    sqlx::query!(
-        "DELETE FROM instance_servers WHERE instance_id = ?",
-        instance_id,
-    )
-    .execute(&mut **tx)
-    .await?;
-    for server in servers {
-        let source = server.source.as_str();
-        let nbt = nbt_to_bytes(&server.data)?;
-        sqlx::query!(
-            "
-			INSERT INTO instance_servers
-				(instance_id, id, source, excluded_synced_server_id,
-				 nbt, position)
-			VALUES (?, ?, ?,
-				(SELECT id FROM synced_servers WHERE id = ?), ?, ?)
-			",
-            instance_id,
-            server.id,
-            source,
-            server.excluded_synced_server_id,
-            nbt,
-            server.position,
-        )
-        .execute(&mut **tx)
-        .await?;
-    }
-    Ok(())
-}
-
-async fn load_projection_entries(
-    instance_id: &str,
-    state: &State,
-) -> crate::Result<Vec<ProjectionEntry>> {
-    let rows = sqlx::query!(
-        "
-		SELECT server_id, owner, nbt, position
-		FROM instance_server_projection_entries
-		WHERE instance_id = ?
-		ORDER BY position
-		",
-        instance_id,
-    )
-    .fetch_all(&state.pool)
-    .await?;
-    rows.into_iter()
-        .map(|row| {
-            Ok(ProjectionEntry {
-                id: row.server_id,
-                owner: ProjectionOwner::from_str(&row.owner).ok_or_else(
-                    || {
-                        ErrorKind::InputError(format!(
-                            "Unknown server projection owner {}",
-                            row.owner
-                        ))
-                    },
-                )?,
-                data: nbt_from_bytes(row.nbt)?,
-                position: row.position,
-            })
-        })
-        .collect()
-}
-
-async fn begin_server_checkpoint(
-    instance_id: &str,
-    servers: &[ServerRecord],
-    expected_sha1: &str,
-    source_revision: i64,
-    state: &State,
-) -> crate::Result<()> {
-    let mut tx = state.pool.begin().await?;
-    sqlx::query!(
-        "DELETE FROM instance_server_projection_entries WHERE instance_id = ?",
-        instance_id,
-    )
-    .execute(&mut *tx)
-    .await?;
-    for (position, server) in servers.iter().enumerate() {
-        let owner = if server.source == ServerSource::UserSynced {
-            ProjectionOwner::Synced
-        } else {
-            ProjectionOwner::Instance
-        }
-        .as_str();
-        let nbt = nbt_to_bytes(&server.data)?;
-        let position = position as i64;
-        sqlx::query!(
-            "
-			INSERT INTO instance_server_projection_entries
-				(instance_id, owner, server_id, nbt, position)
-			VALUES (?, ?, ?, ?, ?)
-			",
-            instance_id,
-            owner,
-            server.id,
-            nbt,
-            position,
-        )
-        .execute(&mut *tx)
-        .await?;
-    }
-    sqlx::query!(
-        "
-		INSERT INTO instance_sync_checkpoints
-			(instance_id, feature, variant, expected_sha1, merge_base,
-			 source_revision, status, link_mode)
-		VALUES (?, 'multiplayer_servers', 'default', ?, NULL, ?,
-			'pending', NULL)
-		ON CONFLICT(instance_id, feature, variant) DO UPDATE SET
-			expected_sha1 = excluded.expected_sha1,
-			merge_base = NULL,
-			source_revision = excluded.source_revision,
-			status = 'pending',
-			link_mode = NULL
-		",
-        instance_id,
-        expected_sha1,
-        source_revision,
-    )
-    .execute(&mut *tx)
-    .await?;
-    tx.commit().await?;
-    Ok(())
-}
-
-fn generated_path(state: &State, instance_id: &str) -> PathBuf {
-    synced_options_path(state)
-        .join("servers/generated")
-        .join(safe_instance_id(instance_id))
-        .join(SERVERS_FILE)
-}
-
-fn is_modpack_link(link: &InstanceLink) -> bool {
-    matches!(
-        link,
-        InstanceLink::ModrinthModpack { .. }
-            | InstanceLink::ImportedModpack { .. }
-            | InstanceLink::SharedInstance { .. }
-    )
-}
-
-fn modpack_version_id(link: &InstanceLink) -> Option<&str> {
-    match link {
-        InstanceLink::ModrinthModpack { version_id, .. } => Some(version_id),
-        InstanceLink::ServerProjectModpack {
-            content_version_id, ..
-        } => Some(content_version_id),
-        InstanceLink::ImportedModpack {
-            version_id: Some(version_id),
-            ..
-        } => Some(version_id),
-        InstanceLink::SharedInstance {
-            modpack_version_id: Some(version_id),
-            ..
-        } => Some(version_id),
-        _ => None,
-    }
 }
